idea for using devcontainer instead of docker-compose
This is a first rough idea on how we could use devcontainers as the testing environment.
The hope was to eliminate the docker-compose.yml and simplify things.
But in the end it solved hardly any problems and made things even slightly more complex.
For the moment I leave it as it is.